Remove collider texture, just use drawing plain old sdl_rects. Create Utils funcs for draing rects and setting render draw color.

+++ sdlgame_win/src/MouseHandler.h
@@ -19,7 +19,7 @@ public:
 	MouseButtonHandler();
 	~MouseButtonHandler() {}
 
-	static const int numButtons = std::numeric_limits<uint8_t>::max();
+	static const uint8_t numButtons = std::numeric_limits<uint8_t>::max();
 
 	void handleMouseButtonEvent(SDL_MouseButtonEvent event);
 	const MouseButtonData& mouseButtonData(uint8_t button);
